+++ src/main/java/stawa/vitalstrike/package-info.java
@@ -29,6 +29,6 @@
  * </ul>
  * 
  * @author Stawa
- * @version 1.1
+ * @version 1.2
  */
 package stawa.vitalstrike;